+++ src/loader.cpp
@@ -57,12 +57,11 @@ void Loader::calculateFunctionSizes() {
 
 void Loader::loadJumpTable(ifstream& in) {
     // load jump table
-    // FIXME: unsigned -> uint64_t
-    unsigned lib_total_jumps;
+    uint64_t lib_total_jumps;
     in.read((char*)&lib_total_jumps, sizeof(decltype(lib_total_jumps)));
 
     uint64_t lib_jmp;
-    for (unsigned i = 0; i < lib_total_jumps; ++i) {
+    for (uint64_t i = 0; i < lib_total_jumps; ++i) {
         in.read((char*)&lib_jmp, sizeof(decltype(lib_jmp)));
         jumps.push_back(lib_jmp);
     }
